requirements.txt
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.13
django-coreapi==2.3.0
django-filter==2.2.0
django-rest-swagger==2.2.0
djangorestframework==3.9.4
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
Markdown==3.0.1
MarkupSafe==1.1.1
numpy==1.16.5
openapi-codec==1.3.2
pandas==0.22.0
psycopg2==2.8.4
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
requests==2.21.0
simplejson==3.17.0
six==1.13.0
uritemplate==3.0.0
urllib3==1.24.3
